+++ ethereum/executionlayer/src/main/java/tech/pegasys/teku/ethereum/executionlayer/ExecutionLayerManagerImpl.java
@@ -13,8 +13,8 @@
 
 package tech.pegasys.teku.ethereum.executionlayer;
 
-import static tech.pegasys.teku.spec.config.Constants.MAXIMUM_CONCURRENT_EB_REQUESTS;
 import static tech.pegasys.teku.spec.config.Constants.MAXIMUM_CONCURRENT_EE_REQUESTS;
+import static tech.pegasys.teku.spec.config.Constants.MAXIMUM_CONCURRENT_NON_CRITICAL_EB_REQUESTS;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.util.Arrays;
@@ -139,7 +139,7 @@ public class ExecutionLayerManagerImpl implements ExecutionLayerManager {
     final MetricRecordingBuilderClient metricRecordingBuilderClient =
         new MetricRecordingBuilderClient(restBuilderClient, timeProvider, metricsSystem);
     return new ThrottlingBuilderClient(
-        metricRecordingBuilderClient, MAXIMUM_CONCURRENT_EB_REQUESTS, metricsSystem);
+        metricRecordingBuilderClient, MAXIMUM_CONCURRENT_NON_CRITICAL_EB_REQUESTS, metricsSystem);
   }
 
   private ExecutionLayerManagerImpl(