+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/datacolumns/db/DasDb.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright Consensys Software Inc., 2024
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.statetransition.datacolumns.db;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import java.util.Optional;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+
+class DasDb extends AbstractDelegatingDasDb {
+
+  private final DataColumnSidecarDB delegate;
+
+  public DasDb(final DataColumnSidecarDB delegate) {
+    super(delegate);
+    this.delegate = checkNotNull(delegate);
+  }
+
+  @Override
+  public SafeFuture<Optional<UInt64>> getFirstCustodyIncompleteSlot() {
+    return delegate.getFirstCustodyIncompleteSlot();
+  }
+
+  @Override
+  public SafeFuture<Optional<UInt64>> getFirstSamplerIncompleteSlot() {
+    return delegate.getFirstSamplerIncompleteSlot();
+  }
+
+  @Override
+  public SafeFuture<Void> setFirstCustodyIncompleteSlot(final UInt64 slot) {
+    return delegate.setFirstCustodyIncompleteSlot(slot);
+  }
+
+  @Override
+  public SafeFuture<Void> setFirstSamplerIncompleteSlot(final UInt64 slot) {
+    return delegate.setFirstSamplerIncompleteSlot(slot);
+  }
+}